Extract RuntimeRules into a runtime_rules module
runtime_optimizer.rs had grown to mix the RTO coordinator, the
RuntimeRule trait, and concrete rules in one file. Split:

  runtime_optimizer.rs         RuntimeRule trait + RuntimeOptimizerExec
                               + CoordinatorStream + helpers
  runtime_rules/mod.rs         module index + re-exports
  runtime_rules/swap_build_side_if_inverted.rs
                               SwapBuildSideIfInverted impl + its
                               helpers (just_completed_stage_of_join,
                               ensure_collect_left_single_partition,
                               side_runtime_rows,
                               sum_runtime_rows_across_partitions)

Pure refactor — 158 lines moved without behavior change. Future rules
land alongside, one file each. The single import-site update at
physical-optimizer/runtime_optimizer.rs splits the existing combined
import into two lines (the trait still lives at
datafusion_physical_plan::runtime_optimizer; the concrete rule moves
to datafusion_physical_plan::runtime_rules).

SLT and clippy unchanged.
